Lazy article content and memoized block parsing
ContentBlockStack re-parsed the full article text with regex on every
body evaluation, and the reader's plain VStack materialized every
block (starting all image loads) the moment an article opened.
