+++ App/Display Styles/Scroll/ScrollExpandedArticleView.swift
@@ -36,7 +36,7 @@ struct ScrollExpandedArticleView: View {
 
     var body: some View {
         ScrollView(.vertical, showsIndicators: false) {
-            VStack(alignment: .leading, spacing: 14) {
+            LazyVStack(alignment: .leading, spacing: 14) {
                 headerSection
 
                 if isExtracting && extractedText == nil {
